Fix all clippy warnings for CI (-D warnings, --all-targets --features mcp)
- Remove unused import MeetingType (commands.rs)
- manual_range_contains: days >= 29 && days <= 31 → (29..=31).contains (intelligence.rs)
- unnecessary_cast: WORK_DAY_END_HOUR as u32 → WORK_DAY_END_HOUR (gaps.rs)
- needless_borrows_for_generic_args: &[...] → [...] (extract.rs, mod.rs)
- too_many_arguments: allow on insert_capture + build_entity_result (impact_rollup.rs, mcp.rs)

Verified: cargo clippy --all-targets --features mcp -- -D warnings passes clean.

+++ src-tauri/src/prepare/gaps.rs
@@ -266,7 +266,7 @@ mod tests {
         for gap in &gaps {
             let end_str = gap.get("end").and_then(|v| v.as_str()).unwrap_or("");
             let end_dt = parse_event_dt(end_str, None).unwrap();
-            assert!(end_dt.hour() <= WORK_DAY_END_HOUR as u32);
+            assert!(end_dt.hour() <= WORK_DAY_END_HOUR);
         }
     }
